refactor(ds): make ModalPortal component
